% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{cf}
\alias{cf}
\alias{gpcf_const}
\alias{gpcf_lin}
\alias{gpcf_sexp}
\alias{gpcf_matern32}
\alias{gpcf_matern52}
\title{Initialize covariance function}
\usage{
gpcf_const(magn = 1)

gpcf_lin(ind = NULL, magn = 1)

gpcf_sexp(ind = NULL, lscale = 0.5, magn = 1)

gpcf_matern32(ind = NULL, lscale = 0.5, magn = 1)

gpcf_matern52(ind = NULL, lscale = 0.5, magn = 1)
}
\arguments{
\item{magn}{Initial value for the magnitude hyperparameter (depicts the magnitude of the variation captured by the given covariance function).}

\item{ind}{Indices of the inputs which are taken into account when calculating this covariance. Default is all the inputs.}

\item{lscale}{Initial value for the length-scale hyperparameter.}
}
\value{
The covariance function object.
}
\description{
Functions for initializing the covariance functions which can then be passed to \code{\link{gp_init}}.
}
\examples{
\donttest{
# Basic usage (single covariance function)
cf <- gpcf_sexp()
lik <- lik_binomial()
gp <- gp_init(cf, lik)
gp <- gp_optim(gp, x ,y, trials)


# More than one covariance function, and the function is additive 
# in the first input, but has an interaction between inputs 2 and 3
# (the functional form is f(x) = f(x_1) + f(x_2,x_3))
cfs <- list(gpcf_const(), gpcf_sexp(1), gpcf_sexp(c(2,3)))
lik <- lik_gaussian()
gp <- gp_init(cf, lik)
gp <- gp_optim(gp, x, y)

}

}
