% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp_energy}
\alias{gp_energy}
\title{Energy of a GP model}
\usage{
gp_energy(gp, include_prior = T)
}
\arguments{
\item{gp}{The fitted GP model.}
}
\value{
The energy value (negative log marginal likelihood).
}
\description{
Returns the energy (negative log marginal likelihood) of a fitted GP model with the current hyperparameters. The result is exact for the Gaussian likelihood and based on Laplace approximation for other cases.
}
\section{References}{


Rasmussen, C. E. and Williams, C. K. I. (2006). Gaussian processes for machine learning. MIT Press.
}

\examples{
\donttest{
# Basic usage (single covariance function)
cf <- cf_sexp()
lik <- lik_binomial()
gp <- gp_init(cf, lik)
e <- gp_energy(gp)

}

}
