% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_optim.R
\name{gp_optim}
\alias{gp_optim}
\title{Optimize hyperparameters of a GP model}
\usage{
gp_optim(gp, x, y, method = "Nelder-Mead", tol = 1e-04, verbose = T,
  ...)
}
\arguments{
\item{gp}{The gp model object to be fitted.}

\item{x}{n-by-d matrix of input values (n is the number of observations and d the input dimension). 
Can also be a vector of length n if the model has only a single input.}

\item{y}{Vector of n output (target) values.}

\item{method}{Optimization method that will be passed to \link{optim} function.}

\item{tol}{Relative change in the objective function value below the optimization is terminated.}

\item{verbose}{If TRUE, then some information about the progress of the optimization is printed to the console.}

\item{...}{Further arguments to be passed to \link{gp_fit} that are needed in the fitting process, for
example \code{trials} in the case of binomial likelihood.}
}
\value{
An updated GP model object.
}
\description{
This function can be used to optimize the hyperparameters of the model to the maximum marginal
likelihood solution (type-II maximum likelihood) based on the Laplace approximation.
}
\section{References}{


Rasmussen, C. E. and Williams, C. K. I. (2006). Gaussian processes for machine learning. MIT Press.
}

\examples{
\donttest{
# Basic usage (single covariance function)
cf <- cf_sexp()
lik <- lik_binomial()
gp <- gp_init(cf, lik)
gp <- gp_optim(gp, x, y, trials=trials)

}



}
