% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_pred.R
\name{gp_pred}
\alias{gp_pred}
\title{Make predictions with a GP model}
\usage{
gp_pred(gp, xt, var = F, draws = NULL, transform = T,
  jitter = NULL)
}
\arguments{
\item{gp}{A fitted GP model object.}

\item{xt}{N-by-d matrix of input values (N is the number of test points and d the input dimension). 
Can also be a vector of length N if the model has only a single input.}

\item{var}{Whether to compute the predictive variances along with predictive mean}

\item{draws}{Number of draws to generate from the predictive distribution for the latent values.}

\item{transform}{Whether to transform the draws of latent values to the same scale as the target y.}

\item{jitter}{Magnitude of diagonal jitter for covariance matrices for numerical stability. Default is 1e-4 for Gaussian and 1e-2 for other likelihoods.}
}
\value{
If draws is given (sampling from the predictive distribution), then returns an N-by-draws
matrix of random draws from the predictive distribution. If draws is not set, then returns either only
the predictive means (if \code{var=F}) or a list with both the predictive means and variances (if \code{var=T}).
}
\description{
This function can be used to make predictions with a fitted model.
}
\section{References}{


Rasmussen, C. E. and Williams, C. K. I. (2006). Gaussian processes for machine learning. MIT Press.
}

\examples{
\donttest{
# fit GP
gp <- gp_init()
gp <- gp_optim(gp, x, y, trials=trials)

# analytic prediction
pred <- gp_pred(gp, xt, var=T)
pred$mean 
pred$var 

# draw from the predictive distribution based on the analytic posterior approximation
draws <- gp_pred(gp, xt, draws=1000) 

# fit using mcmc and predict using that
gpmc <- gp_sample(gp, x, y, trials=trials, iter=1000, chains=2)
draws <- gp_pred(gpmc, xt, draws=1000)
}


}
