% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplite-package.R
\docType{package}
\name{gplite-package}
\alias{gplite-package}
\alias{gplite}
\title{The 'gplite' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. https://mc-stan.org
}
