% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik.R
\name{lik}
\alias{lik}
\alias{lik_gaussian}
\alias{lik_binomial}
\alias{lik_betabinom}
\title{Initialize likelihood}
\usage{
lik_gaussian(sigma = 0.5)

lik_binomial(link = "logit")

lik_betabinom(link = "logit", phi = 0.1)
}
\arguments{
\item{sigma}{Initial value for the noise standard deviation.}

\item{link}{Link function if the likelihood supports non-identity links.}
}
\value{
The likelihood object.
}
\description{
Functions for initializing the likelihood (observations model) which can then be passed to \code{\link{gp_init}}.
}
\details{
Different likelihoods have the following possible link functions:
\describe{
 \item{\code{lik_gaussian}}{No links (uses identity link).}
 \item{\code{lik_binomial}}{'logit' or 'probit'}
}
}
\examples{
\donttest{
# Basic usage 
cf <- cf_sexp()
lik <- lik_binomial()
gp <- gp_init(cf, lik)
gp <- gp_optim(gp, x ,y, trials)

}

}
