% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior}
\alias{prior}
\alias{prior_fixed}
\alias{prior_logunif}
\title{Initialize prior for hyperparameter}
\usage{
prior_fixed()

prior_logunif()
}
\value{
The hyperprior object.
}
\description{
Functions for initializing hyperparameter priors which can then be passed 
to \code{\link{gp_init}}. See section Details for the prior explanations.
}
\details{
The supported priors are:
\describe{
 \item{\code{prior_fixed}}{ The hyperparameter is fixed to its initial value, and is not optimized by 
                            \code{gp_optim}. }
 \item{\code{prior_logunif}}{ Improper uniform prior on the log of the parameter. }
}
}
\section{References}{


Rasmussen, C. E. and Williams, C. K. I. (2006). Gaussian processes for machine learning. MIT Press.
}

\examples{
\donttest{
# Fix one of the hyperparameters
cf <- cf_nn(sigma0 = 0, prior_sigma0 = prior_fixed())
lik <- lik_gaussian()
gp <- gp_init(cf, lik)
gp <- gp_optim(gp, x, y)


}

}
