% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp_fit}
\alias{gp_fit}
\title{Fit a GP model}
\usage{
gp_fit(gp, x, y, trials = NULL, jitter = 0.001, ...)
}
\arguments{
\item{gp}{The gp model object to be fitted.}

\item{x}{n-by-d matrix of input values (n is the number of observations and d the input dimension). Can also be a vector of length n if the model has only a single input.}

\item{y}{vector of n output (target) values.}
}
\value{
An updated GP model object that can be passed to other functions, for example when optimizing the hyperparameters or making predictions.
}
\description{
Fits a GP model with the current hyperparameters. Notice that this function does not optimize the hyperparameters in any way, but only finds the Laplace approximation (or the analytical true posterior in the case of Gaussian likelihood) to the latent values.
}
\section{References}{


Rasmussen, C. E. and Williams, C. K. I. (2006). Gaussian processes for machine learning. MIT Press.
}

\examples{
\donttest{
### Basic usage (single covariance function)
cf <- gpcf_sexp()
lik <- lik_binomial()
gp <- gp_init(cf, lik)
gp <- gp_optim(gp, x ,y, trials)

}

}
