% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp_init}
\alias{gp_init}
\title{Initialize a GP model}
\usage{
gp_init(cfs = gpcf_sexp(), lik = lik_gaussian())
}
\arguments{
\item{lik}{Likelihood (observation model). See \code{\link{lik}}.}

\item{cf}{The covariance function(s). Either a single covariance function or a list of them. See \code{\link{cf}}.}
}
\value{
A GP model object that can be passed to other functions, for example when optimizing the hyperparameters or making predictions.
}
\description{
Initializes a GP model with given covariance function(s) and likelihood. The model can then be fitted using \code{\link{gp_fit}} or \code{\link{gp_sample}}. For hyperparameter optimization, see \code{\link{gp_optim}}
}
\section{References}{


Rasmussen, C. E. and Williams, C. K. I. (2006). Gaussian processes for machine learning. MIT Press.
}

\examples{
\donttest{
### Basic usage (single covariance function)
cf <- gpcf_sexp()
lik <- lik_binomial()
gp <- gp_init(cf, lik)
gp <- gp_optim(gp, x ,y, trials)

}

}
