% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp_init}
\alias{gp_init}
\title{Initialize a GP model}
\usage{
gp_init(
  cfs = cf_sexp(),
  lik = lik_gaussian(),
  method = method_full(),
  approx = approx_laplace()
)
}
\arguments{
\item{cfs}{The covariance function(s). Either a single covariance function or a list of them. See \code{\link{cf}}.}

\item{lik}{Likelihood (observation model). See \code{\link{lik}}.}

\item{method}{Method for approximating the covariance function.
See \code{\link{method}}.}

\item{approx}{Approximate inference method for Gaussian approximation
for the posterior of the latent values. See \code{\link{approx}}.}
}
\value{
A GP model object that can be passed to other functions, for example when optimizing the hyperparameters or making predictions.
}
\description{
Initializes a GP model with given covariance function(s) and likelihood. The model can then be fitted using \code{\link{gp_fit}}. For hyperparameter optimization, see \code{\link{gp_optim}}
}
\section{References}{


Rasmussen, C. E. and Williams, C. K. I. (2006). Gaussian processes for machine learning.
MIT Press.
}

\examples{
\donttest{

# Generate some toy data
set.seed(1242)
n <- 500
x <- matrix(rnorm(n * 3), nrow = n)
f <- sin(x[, 1]) + 0.5 * x[, 2]^2 + x[, 3]
y <- f + 0.5 * rnorm(n)
x <- data.frame(x1 = x[, 1], x2 = x[, 2], x3 = x[, 3])

# Full exact GP with Gaussian likelihood
gp <- gp_init(cf_sexp())
gp <- gp_optim(gp, x, y)

# Approximate solution using random features
gp <- gp_init(cf_sexp(), method = method_rf(num_basis = 300))
gp <- gp_optim(gp, x, y)

# Approximate solution using FITC
gp <- gp_init(cf_sexp(), method = method_fitc(num_inducing = 100))
gp <- gp_optim(gp, x, y)
}

}
