% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik.R
\name{lik}
\alias{lik}
\alias{lik_gaussian}
\alias{lik_binomial}
\alias{lik_betabinom}
\title{Initialize likelihood}
\usage{
lik_gaussian(sigma = 0.5, prior_sigma = prior_logunif())

lik_binomial(link = "logit")

lik_betabinom(link = "logit", phi = 0.1, prior_phi = prior_logunif())
}
\arguments{
\item{sigma}{Initial value for the noise standard deviation.}

\item{link}{Link function if the likelihood supports non-identity links. See Details for 
information about possible links for each likelihood.}

\item{phi}{The over dispersion parameter for beta binomial likelihood.}
}
\value{
The likelihood object.
}
\description{
Functions for initializing the likelihood (observation model) which can then be passed to \code{\link{gp_init}}.
}
\details{
The supported likelihoods are:
\describe{
 \item{\code{lik_gaussian}}{Gaussian likelihood. Has no links (uses identity link).}
 \item{\code{lik_binomial}}{Binomial likelihood. Possible links: 'logit' or 'probit'.}
 \item{\code{lik_betabinom}}{Beta binomial likelihood. Possible links: 'logit' or 'probit'.}
}
}
\examples{
\donttest{
# Basic usage 
cf <- cf_sexp()
lik <- lik_binomial()
gp <- gp_init(cf, lik)
gp <- gp_optim(gp, x ,y, trials)

}

}
